Refine logic for placing points on map during test and how they are stored, also many schema/class changes.
Too many changes to document for db_schema_classes and test_class_implementations since they are basically totally new anyway.
The gist of it is that there is an abstract class Test acting as a parent for all future test classes, including the already mostly implemented LightingProfileTest class.
Test has many methods that need to be implemented and requirements for constructors but there are a ton of documentation comments in there so just look at those.