Add function Test.createTestPageFor and rename project_details
New function in Test createTestPageFor abstracts away even more of the process for figuring out the specific type and page for a given Test object. So ideally you only need the Test and of course the Project it came from and then using createTestPageFor gives you the page to complete that test with no extra work on project_details_page or wherever else needs that functionality.
create_project_details.dart is now project_details_page.dart