Rework Test class and the way it connects to subclasses one final time.
After this rework there is no longer a need to look inside the Test class and add switch cases for each new subclass.
Instead, each new subclass must register themselves in the Maps used by Test to determine which constructor or widget to use based on collection ID or object type. Preferably this is done in a single-use method such as register() seen in LightingProfileTest, and then all of these are called once when first launching the app to populate the Maps in Test class.