Fix/adjust various things around the people tests.
Creating people tests works now, the main problem was just not having created rules for accessing those collection names in Firestore yet, but that is done now.
The tests both navigate to their page correctly but beyond that have varying levels of functionality.
However, mostly it seems like there is not a button setup to submit the data when you're done and I'm not quite sure what you intended for that so I did not change it. Similar case for saveTracedRoute function in firestore_functions.
Lines added to pubspec.yaml for deeper asset directories used in people tests were necessary for the marker icons to work (at least for me).
I also did a decent bit of adjusting function/constructor signatures such as on the test pages to be simpler and standardized with the others as simply taking the Project and Test objects is all that is necessary since all the data previously being passed through other parameters is contained in those 2 objects.