Finish spatial boundaries test functionality. Placement of all boundary types now works as well as submitting data to Firestore.
Reworked much of the structure of classes for spatial boundaries in db_schema_classes once again in the process of finishing everything.
Also removed a print statement in nature prevalence recreateTestConstructors.
Add extension functions getAreaInSquareFeet for Polygon and getLengthInFeet for Polyline, as well as reorganize the extensions for those two types in firestore_functions.
Adjust finalizePolygon and createPolyline in google_maps_functions to create a copy of the list of points they are given rather than using the reference to remove potential issues down the line caused by using reference.
Remove unnecessary imports and variables from lighting_profile_test.
Update gradle and android versions in settings.gradle and gradle-wrapper.properties.