Finish backend for completing people in place test.
This does not include the addition of standing point or timer functionality, just the basic ability to submit data as most other tests have at this point.
The 'Start' button for the timer is enabled and changes to say 'End' which allows you to end the test and submit data, but the timer does not seem to work properly and there is still no backend support for a timer.
This update includes lots of tweaks and refactors to various parts of the people in place test and related classes in db_schema_classes.
The most noticeable of these other changes is the gender options being changed to include 'nonbinary' and 'unspecified', while 'N/A' was removed for just not making sense as an option.
Internally the biggest change was the refactoring of basically everything around the modal sheet for giving a description of the person. The way data is used in and around the widget is completely different, and all the buttons previously assembled through a few build methods are now ChoiceChips or FilterChips, all of which are defined in a clear way within the build method itself rather than packed away in a series of methods which also contained the text/spacing around them, which made it difficult to understand what was happening from the code imo.