Refactor some of SectionCutterTest and Section class to use JsonToString mixin with them.
Shouldn't have done this here or now probably but oh well. I just wanted to mix in JsonToString with every Test subclass and remove toString override in Test class itself, but forgot that refactoring needs to be done for some of them in order to achieve that and was in too deep on the refactor to stop.