Update layout and add timer backend to People in Motion.
Also changed how icons relating to the activity types are handled in People in Motion. Added the file path to each icon as a string on the enum itself. Then a map from the enum ActivityTypeInMotion to an AssetMapBitmap that is constructed based on that type's iconName value can be easily constructed. This is now how the icons are constructed in assets.dart and used on people_in_motion_test.dart.