Fix selectedTeam logic again and tweaks to deletion.
Found some seemingly inconsistent issues with deleting nested elements of teams/projects, like entire projects and all their content within a team not being deleted when the team was but with no apparent reason and it doesn't happen every time. Tried to fix it but not sure I did anything really.